requirements.txt
Django==2.2
django-tinymce==2.8.0
mysqlclient==1.4.6
Pillow==6.2.1
xlrd==1.2.0
openpyxl==3.0.5
transliterate==1.10.2
django-debug-toolbar==2.2
django-debug-toolbar-template-timings==0.9